move common.py
