+++ brain_games/common.py
